app/user/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
grpc-interceptor ~= 0.11.0
grpcio-tools ~= 1.30